+++ exchanges/kraken/kraken_types.go
@@ -803,6 +803,59 @@ func (e errorResponse) Warnings() string {
 	return strings.Join(e.warnings, ", ")
 }
 
+type wsTicker struct {
+	Ask                        [3]types.Number `json:"a"`
+	Bid                        [3]types.Number `json:"b"`
+	Last                       [2]types.Number `json:"c"`
+	Volume                     [2]types.Number `json:"v"`
+	VolumeWeightedAveragePrice [2]types.Number `json:"p"`
+	Trades                     [2]int64        `json:"t"`
+	Low                        [2]types.Number `json:"l"`
+	High                       [2]types.Number `json:"h"`
+	Open                       [2]types.Number `json:"o"`
+}
+
+type wsSpread struct {
+	Bid       types.Number
+	Ask       types.Number
+	Time      types.Time
+	BidVolume types.Number
+	AskVolume types.Number
+}
+
+func (w *wsSpread) UnmarshalJSON(data []byte) error {
+	return json.Unmarshal(data, &[5]any{&w.Bid, &w.Ask, &w.Time, &w.BidVolume, &w.AskVolume})
+}
+
+type wsTrades struct {
+	Price     types.Number
+	Volume    types.Number
+	Time      types.Time
+	Side      string
+	OrderType string
+	Misc      string
+}
+
+func (w *wsTrades) UnmarshalJSON(data []byte) error {
+	return json.Unmarshal(data, &[6]any{&w.Price, &w.Volume, &w.Time, &w.Side, &w.OrderType, &w.Misc})
+}
+
+type wsCandle struct {
+	LastUpdateTime types.Time
+	EndTime        types.Time
+	Open           types.Number
+	High           types.Number
+	Low            types.Number
+	Close          types.Number
+	VWAP           types.Number
+	Volume         types.Number
+	Count          int64
+}
+
+func (w *wsCandle) UnmarshalJSON(data []byte) error {
+	return json.Unmarshal(data, &[9]any{&w.LastUpdateTime, &w.EndTime, &w.Open, &w.High, &w.Low, &w.Close, &w.VWAP, &w.Volume, &w.Count})
+}
+
 type wsSnapshot struct {
 	Asks []wsOrderbookItem `json:"as"`
 	Bids []wsOrderbookItem `json:"bs"`
